# app/backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a modern RSS reader API.",
        "title": "Krss API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/mcp/oauth/callback": {
            "get": {
                "tags": [
                    "mcp"
                ],
                "summary": "OAuth callback",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization code",
                        "name": "code",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "State",
                        "name": "state",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "HTML result page",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/mcp/outbound": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mcp"
                ],
                "summary": "MCP outbound status",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/krss_backend_internal_service.MCPOutboundStatus"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mcp"
                ],
                "summary": "Update MCP outbound options",
                "parameters": [
                    {
                        "description": "Options",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.mcpOutboundOptionsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/krss_backend_internal_service.MCPOutboundStatus"
                        }
                    }
                }
            }
        },
...
        "/mcp/servers": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mcp"
                ],
                "summary": "List MCP servers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_handler.mcpServerResponse"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mcp"
                ],
                "summary": "Create a MCP server",
                "parameters": [
                    {
                        "description": "MCP server",
                        "name": "server",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.mcpServerRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.mcpServerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.errorResponse"
                        }
                    }
                }
            }
        },
        "/mcp/servers/{id}": {
            "delete": {
                "tags": [
                    "mcp"
                ],
                "summary": "Delete a MCP server",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "MCP server ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            },
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mcp"
                ],
                "summary": "Update a MCP server",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "MCP server ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "MCP server",
                        "name": "server",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.mcpServerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.mcpServerResponse"
                        }
                    }
                }
            }
        },
        "/mcp/servers/{id}/inspect": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mcp"
                ],
                "summary": "Inspect a MCP server (dry run + preview)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "MCP server ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Inspect request",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.mcpInspectRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/krss_backend_internal_service.MCPInspectResult"
                        }
                    }
                }
            }
        },
        "/mcp/servers/{id}/oauth/discovery": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mcp"
                ],
                "summary": "Discover OAuth authorization server",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "MCP server ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/krss_backend_internal_service.OAuthDiscoveryResult"
                        }
                    }
                }
            }
        },
        "/mcp/servers/{id}/oauth/revoke": {
            "post": {
                "tags": [
                    "mcp"
                ],
                "summary": "Revoke OAuth authorization",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "MCP server ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_handler.errorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "internal_handler.mcpOutboundOptionsRequest": {
            "type": "object",
            "properties": {
                "baseUrl": {
                    "description": "BaseURL 对外访问地址（用户填的公网/局域网可达 origin；空字符串 = 清掉）。",
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "writeEnabled": {
                    "type": "boolean"
                }
            }
        },
...
        "krss_backend_internal_service.MCPOutboundStatus": {
            "type": "object",
            "properties": {
                "baseUrl": {
                    "description": "BaseURL 对外访问地址（用户填；空 = 没填，示例里回落到浏览器当前 origin）。",
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "endpoint": {
                    "type": "string"
                },
                "protocolVersion": {
                    "type": "string"
                },
                "resourceCount": {
                    "type": "integer"
                },
                "token": {
                    "type": "string"
                },
                "tokenCreatedAt": {
                    "type": "string"
                },
                "tokenPrefix": {
                    "type": "string"
                },
                "tokenSet": {
                    "description": "TokenSet 是否有可用 Token；Token 是明文库里那份原文（可重复查看复制，17-2）。\n存量哈希没有明文可回 → Token 为空，界面提示点一次「重新生成」。",
                    "type": "boolean"
                },
                "toolCount": {
                    "type": "integer"
                },
                "writeEnabled": {
                    "description": "WriteEnabled 写操作总开关（默认关；第一版只读）。",
                    "type": "boolean"
                }
            }
        },
...
    }
}